requirements.txt
rich>=13.0.0
sentence-transformers>=2.2.0
numpy>=1.24.0
python-dotenv>=1.0.0
openai>=1.0.0
